Create Fan Art model
This creates the 'model', which is what tells Rails which database table to
look at for the FanArt model, so any interactions with that table go through
this file.

This is where you add methods on the object, e.g.

```
def publish
  published = true
end
```

Would let you do `FanArt.last.publish`, which would set the published column
on the FanArt table to true (if it existed).

You can also set 'relations' between this model and others, e.g. 'belongs_to
:user' to give it FanArt.last.user and it actually be an object, not a piece
of text. So you could do FanArt.last.user.name to get the name of the user
that created that FanArt.